flask_app/requirements.txt
flask==2.1.1
celery==5.2.6
paramiko==2.10.3
python-dotenv==0.20.0
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60


